tomato_cloud/requirements.txt
Django==1.5.12
django-filter==0.12.0
djangorestframework==3.2.3
gunicorn==19.4.5
Markdown==2.6.5
wheel==0.29.0
whitenoise==2.0.6
